[PQ-511] renaming from intergation_test to test and fdr deleting
