+++ .gitignore
@@ -12,7 +12,7 @@ __pycache__/
 # Distribution / packaging
 .Python
 build/
-/src/integration_test/gpd/node_modules
+/src/integration/gpd/node_modules
 develop-eggs/
 dist/
 downloads/
