+++ behave.ini
@@ -1,5 +1,5 @@
 [behave]
-paths = src/integration_test
+paths = src/integration
 format=json
 logging_clear_handlers=yes
 logging_filter=-suds
