+++ src/integration/wisp/environment.py
@@ -4,8 +4,8 @@ import logging
 from src.conf.configuration import commondata
 from src.conf.configuration import secrets
 from src.conf.configuration import settings
-from src.integration_test.utility.wisp import constants
-from src.integration_test.utility.wisp.constants import empty_flow_data
+from src.integration.utility.wisp import constants
+from src.integration.utility.wisp.constants import empty_flow_data
 
 
 def before_all(context):
