--- src/integration_test/fdr/steps/helloworld.py
@@ -1,21 +1,0 @@
-from behave import *
-
-
-@given('we have behave installed')
-def step_impl(context):
-    pass
-
-
-@when('we implement a test')
-def step_impl(context):
-    assert True is not False
-
-
-@then('behave will test it for us!')
-def step_impl(context):
-    assert context.failed is False
-
-
-@when('we implement a test with assert false')
-def step_impl(context):
-    assert False
